// src/constants/status.d.ts
export declare enum StatusEnum {
    ON_TIME = "on_time",
    COMPLETED = "completed",
    FAILED = "failed",
    NOT_ARRIVED = "not_arrived",
    AT_RISK = "at_risk",
    UPCOMING = "upcoming",
    EMPTY_STATUS = "",
    CUSTOMER_INFORMED = "Customer Informed",
    TRANSPORTER_LOCKED = "Transporter Locked",
    IDENTITY_VERIFIED = "Identity Verified",
    BOARDING_PASS_VERIFIED = "Boarding pass verified",
    BAGGAGE_CONFIRMED = "Baggage confirmed",
    CUSTOMER_CONFIRMED = "Customer confirmed"
}